+++ src/components/Item.jsx
@@ -1,46 +1,20 @@
 import { useState, useEffect } from "react";
 import { useSelector, useDispatch } from "react-redux";
 import { addBookmark, deleteBookmark } from "../redux/Bookmark-Reducer";
+import useManageBookmarkList from "../hooks/useManageBookmarkList";
+
 import Modal from "./Modal";
 import ItemViewerBox from "./ItemViewerBox";
 
 
 function Item ({ item }) {
 
+    const { bookmark, bookmarkStateChange } = useManageBookmarkList(item);
     const [modalState, setModalState] = useState(false);
 
     const modalStateChange = () => {
         setModalState(!modalState);
     }
-
-    const dispatch = useDispatch();
-
-    // 2) Bookmark 관련 상태
-    const bookmarkList = useSelector(state => state.bookmarkList);
-    let previousBookmarkAdded = bookmarkList.find(bookmarkItem => bookmarkItem.id === item.id); 
-    
-    const [bookmark, setBookmark] = useState(previousBookmarkAdded !== undefined ? true : false); 
-    
-    const bookmarkStateChange = (state) => {
-        setBookmark(state);
-    }
-
-    useEffect(() => { 
-
-        (bookmark === true && previousBookmarkAdded === undefined) && dispatch(addBookmark(item));
-        (bookmark === false) && dispatch(deleteBookmark(item));
-
-    }, [bookmark])
-
-
-    useEffect(() => {
-
-        localStorage.setItem('bookmark', JSON.stringify(bookmarkList)); 
-        (previousBookmarkAdded === 'no') && bookmarkStateChange(false); // 북마크 관련 컴포넌트에서 북마크 해제 -> 상품 리스트 컴포넌트에도 연동
-
-    }, [bookmarkList])
-
-    
 
 
     if(modalState === true) {
